# documents/openapi.yaml
openapi: 3.0.3
info:
  title: HotelBook API
  version: 1.0.0
  description: |
    REST API suy ra từ các màn hình frontend (`apps/web`) và tài liệu
    [`screens.md`](./screens.md), map với ERD [`sql/erd.md`](./sql/erd.md).

    **Quy ước**
    - Prefix `/api/v1`
    - JWT Bearer tách biệt: khách (`USERS`) vs admin (`ADMINS`)
    - ID tài nguyên là `integer` theo schema; mã hiển thị (`DP-0002`, `RF-01`)
      do server format
    - Giỏ đặt phòng (P4) giữ phía client; checkout gửi `phong_ids` + ngày + method
    - `description` / `image` trên loại phòng có trên UI nhưng chưa có trong ERD
      — schema đánh dấu `nullable` để frontend không gãy khi backend chưa lưu

    Mở file này trên https://editor.swagger.io để xem Swagger UI.
servers:
  - url: http://localhost:8001
    description: Local (docker-compose / FastAPI)
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token sau POST /api/v1/auth/login hoặc /register
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token sau POST /api/v1/admin/auth/login
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        phone:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, LOCKED]
...
    AdminAccount:
      type: object
      properties:
        id:
          type: integer
        full_name:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, STAFF]
        status:
          type: string
          enum: [ACTIVE, LOCKED]
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký khách hàng
      description: P5 — tab Đăng ký. Tạo `USERS` rồi trả JWT.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "409":
          description: Email đã tồn tại
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập khách hàng
      description: P5 — tab Đăng nhập
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Sai thông tin đăng nhập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Đăng xuất khách hàng
      description: Header public — Đăng xuất
      operationId: logoutUser
      security:
        - UserBearer: []
      responses:
        "204":
          description: No Content
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Thông tin khách đang đăng nhập
      description: Header (tên user). P10 chưa có UI — có thể dùng sau để sửa profile.
      operationId: getCurrentUser
      security:
        - UserBearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /api/v1/admin/auth/login:
    post:
      tags: [AdminAuth]
      summary: Đăng nhập admin
      description: A1 — bảng `ADMINS`, tách biệt `USERS`
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Sai thông tin đăng nhập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/auth/logout:
    post:
      tags: [AdminAuth]
      summary: Đăng xuất admin
      operationId: logoutAdmin
      security:
        - AdminBearer: []
      responses:
        "204":
          description: No Content
  /api/v1/admin/auth/me:
    get:
      tags: [AdminAuth]
      summary: Admin đang đăng nhập
      description: A2 sidebar — tên / vai trò
      operationId: getCurrentAdmin
      security:
        - AdminBearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
...